% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Description_Fake_COVARS.R
\docType{data}
\name{Fake_COVARS}
\alias{Fake_COVARS}
\title{5,000 Participant Simulated Sample of GWAS-like Covariate information}
\format{
A data frame with 5,000 rows and 8 variables (simulated)
}
\usage{
data(Fake_COVARS)
}
\description{
A dataset containing simulated quantitative, binary and categorical covariates for participants
The variables are as follows:
}
\details{
\itemize{
\item FID = Participant Family Identifier
\item IID = Participant Individual Identifier
\item Age = Years of Age
\item Sex = Genetic Sex
\item Location = Sample Region
\item BMI = Continous body mass index
\item PC1 = First Genetic Principle Component
\item PC2 = Second Genetic Principle Component

}
}
\keyword{datasets}
