% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fake_Demo_Data.R
\docType{data}
\name{Fake_Demo_Data}
\alias{Fake_Demo_Data}
\title{1,000 Participant Simulated Sample of Dementia Cases, with Covariate information for modelling}
\format{
A data frame with 1,000 rows and 6 variables (simulated)
}
\usage{
data(Fake_Demo_Data)
}
\description{
A dataset containing simulated dementia cases and covariates for participants
The variables are as follows:
}
\details{
\itemize{
\item Age = Age (in years)
\item Sex = Genetic Sex
\item BMI = Continous body mass index
\item Ethnicity = Broad self-report ethnicity
\item Dementia = Diagnosis of dementia, with 1 representing a case
\item Location = Sample site, grouped by dwelling type

}
}
\keyword{datasets}
