% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Description_Fake_HLA_10K.R
\docType{data}
\name{Fake_HLA_10K}
\alias{Fake_HLA_10K}
\title{10,000 Result Sample (simulated) Subset of HLA-WAS Summary Statistics adapted from Intelligence_Sum_Stats}
\format{
A data frame with 10,000 rows (randomly selected) and 6 variables
}
\usage{
data(Fake_HLA_10K)
}
\description{
A dataset containing a simulated HLA adaptation of the summary statistics from the Savage Jansen 2018 Intelligence GWAS
}
\details{
The variables are as follows:

\itemize{

\item SNP: HLA identifier
\item CHR: chromosome number
\item POS: base pair position reported on GRCh37
\item A1: effect allele
\item A2: non-effect allele
\item P: P-value

}
}
\keyword{datasets}
