% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Description_Fake_PHENOS_Binary.R
\docType{data}
\name{Fake_PHENOS_Binary}
\alias{Fake_PHENOS_Binary}
\title{5,000 Participant Simulated Sample of Binary Phenotypes}
\format{
A data frame with 5,000 rows and 3 variables (simulated)
}
\usage{
data(Fake_PHENOS_Binary)
}
\description{
A dataset containing simulated binary phenotypes for participants
The variables are as follows:
}
\details{
\itemize{
\item FID = Participant Family Identifier
\item IID = Participant Individual Identifier
\item PHENO_BIN = Binary Outcome

}
}
\keyword{datasets}
