% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Description_Fake_PHENOS_Quant.R
\docType{data}
\name{Fake_PHENOS_Quant}
\alias{Fake_PHENOS_Quant}
\title{5,000 Participant Simulated Sample of Quantitative Phenotypes}
\format{
A data frame with 5,000 rows and 3 variables (simulated)
}
\usage{
data(Fake_PHENOS_Quant)
}
\description{
A dataset containing simulated quantitative phenotypes for participants
The variables are as follows:
}
\details{
\itemize{
\item FID = Participant Family Identifier
\item IID = Participant Individual Identifier
\item PHENO_QUANT = Quantitative Outcome

}
}
\keyword{datasets}
