% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_Plot.R
\name{Forest_Plot}
\alias{Forest_Plot}
\title{Title Create forest plots of test statistics of regression model covariates and GWAS summary statistics for SNPs of interest}
\usage{
Forest_Plot(
  Data_Sets = c("ModelSum", "ModelSum"),
  Names = NULL,
  Data_Set_Colours = viridis::viridis(length(Data_Sets)),
  Chromosome_Columns = c(),
  Left_Spaces = 2,
  Right_Spaces = 2,
  P_Stat_Spaces = 3,
  Model_Reference = FALSE,
  X_Axis_Text_Size = 10,
  Line_Space = 0.01,
  Test_Statistic = NULL,
  Display_Test_Stat_Se_Column = TRUE,
  Display_Test_Stat_CI_Column = FALSE,
  Display_P_Value_Column = TRUE,
  Shapes = NULL,
  Null_Line_Colour = "red",
  Null_Line_Type = "dashed",
  X_Axis_Title = NULL,
  Null_Buffer = 0,
  X_Axis_Title_Size = 15,
  SNP_Stat_Text_Size = 12,
  X_Axis_Label = TRUE,
  X_Axis_Separation = NULL,
  Strip_Colour = "lightblue",
  Strips = TRUE,
  X_Axis_Text_Resolution = 2,
  Pre_Calculated_CIs = FALSE,
  Legend_On = TRUE,
  Legend_Title_Size = 15,
  Legend_Text_Size = 12.5,
  Legend_Title = "Study",
  Left_Title = "SNP",
  P_Value_Title = "p-value",
  Test_Stat_Se_Title = "BETA (SE)",
  OR_Columns = c(),
  Position_Columns = c(),
  SNP_ID_Columns = c(),
  Beta_Columns = c(),
  Standard_Error_Columns = c(),
  PValue_Columns = c(),
  Match_Allele_Direction = TRUE,
  Match_Allele_Study = NULL,
  Selected_SNPs = c(),
  Selected_Covariates = c(),
  Reference_Alleles = c(),
  Effect_Alleles = c(),
  Upper_CI_Columns = c(),
  Lower_CI_Columns = c(),
  File_Name = "Forest_Plot",
  Width = 10,
  Height = 6,
  Quality = 600,
  File_Type = "jpg"
)
}
\arguments{
\item{Data_Sets}{These are a list of GWAS summary statistics or munged model test statistics to be plotted; defaults to c("ModelSum", "ModelSum")}

\item{Names}{These are a list of names for the data sets to be labelled with in the forest plot - this also specifies the order down the page; defaults to c("Vals 1", "Vals 2")}

\item{Data_Set_Colours}{These are a list of colours for the data sets to be coloured with in the forest plot - this also specifies the order down the page; defaults to c("blue", "darkgreen")}

\item{Chromosome_Columns}{These are a list of manual chromosome column names for the data sets used in the order specified; defaults to c("chromosome", "chromosome")}

\item{Model_Reference}{Do you want to use statistics from an R regression model (T/F); defaults to T}

\item{X_Axis_Text_Size}{Size of the X axis text labels; defaults to 15}

\item{Line_Space}{A continuous value specifying the boundary between the plot margin and the left and right lines; defaults to 1}

\item{Test_Statistic}{The type of values being plotted (OR or BETA); defaults to "OR"}

\item{Display_Test_Stat_Se_Column}{Do you want to display the BETA (SE) raw values on the right hand side (T/F); defaults to F}

\item{Display_Test_Stat_CI_Column}{Do you want to display the OR (CI - LL to UL) raw values on the right hand side (T/F); defaults to T}

\item{Display_P_Value_Column}{Do you want to display the P raw values on the right hand side before the test stat column (T/F); defaults to T}

\item{Shapes}{These are a list of shapes corresponding (in the order of) to the data sets to be plotted; defaults to c("square", "diamond")}

\item{Null_Line_Colour}{Colour of null line; defaults to "red"}

\item{Null_Line_Type}{Type of null line; defaults to "dashed"}

\item{X_Axis_Title}{X axis title; defaults to "BETA"}

\item{Null_Buffer}{Units of space to avoid X axis text labels around the nul point; defaults to 0.1}

\item{X_Axis_Title_Size}{X axis title; defaults to 20}

\item{X_Axis_Label}{Do you want an X Axis Label (T/F); defaults to F}

\item{X_Axis_Separation}{Continuous value of the specific break separation to show on the X axis; defaults to 0.02}

\item{Strip_Colour}{Colour of the strips highlighting the groups of covariates or SNPs under a certain study group; defaults to "lightblue"}

\item{Strips}{Do you want above strips to be displayed (T/F); defaults to T}

\item{X_Axis_Text_Resolution}{Number of decimal places to display on X axis text; defaults to 1}

\item{Pre_Calculated_CIs}{Have you pre-calculated UL and LL for CIs (T/F); defaults to F}

\item{Legend_On}{Do you want to display the legend - TRUE/FALSE; defaults to TRUE}

\item{Legend_Title_Size}{Legend title size; defaults to 15}

\item{Legend_Text_Size}{Legend text size; defaults to 12.5}

\item{Legend_Title}{The name of the title to display next to the legend; defaults to "Study"}

\item{Left_Title}{The name of the title to display on the left hand side SNPs or covariates; defaults to "SNP"}

\item{P_Value_Title}{The name of the title to display above P values on right hand side; defaults to "p-value"}

\item{Test_Stat_Se_Title}{The name of the title to display above raw test statistics values on right hand side; defaults to "BETA (SE)"}

\item{OR_Columns}{These are a list of manual OR column names for the data sets used in the order specified; defaults to c("OR", "OR")}

\item{Position_Columns}{These are a list of manual position column names for the data sets used in the order specified; defaults to c("base_pair_location", "base_pair_location")}

\item{SNP_ID_Columns}{These are a list of manual SNP ID column names for the data sets used in the order specified; defaults to c("variant_id","variant_id")}

\item{Beta_Columns}{These are a list of manual BETA column names for the data sets used in the order specified; defaults to  c("beta", "beta")}

\item{Standard_Error_Columns}{These are a list of manual SE column names for the data sets used in the order specified; defaults to c("SE","SE")}

\item{PValue_Columns}{These are a list of manual P column names for the data sets used in the order specified; defaults to c("P","P")}

\item{Match_Allele_Direction}{Do you want to match the allele directions to avoid flips (T/F); defaults to T}

\item{Match_Allele_Study}{If the above is T, which data sets should be the reference point for effect allele direction; defaults to "LbDementia_Sum_Stats"; THIS SHOULD BE THE NAME YOU ASSIGN}

\item{Selected_SNPs}{A list of SNPs to be shown in the forest plot, automatically selected for from the summary statistics data set; defaults to c("rs2616526", "rs7974838", "rs59867714", "rs9571588", "rs79007041")}

\item{Selected_Covariates}{A list of covariates to be shown in the forest plot, automatically selected for from the munged model data set; defaults to c()}

\item{Reference_Alleles}{These are a list of manual reference allele column names for the data sets used in the order specified; defaults to c("other_allele","other_allele")}

\item{Effect_Alleles}{These are a list of manual effect allele column names for the data sets used in the order specified; defaults to c("effect_allele","effect_allele")}

\item{Upper_CI_Columns}{These are a list of manual CI UL column names for the data sets used in the order specified; defaults to c("UL","UL")}

\item{Lower_CI_Columns}{These are a list of manual CI LL column names for the data sets used in the order specified; defaults to c("LL","LL")}

\item{File_Name}{File name to save plot as; defaults to "Forest_Plot"}

\item{Width}{Width of saved plot; defaults to 10}

\item{Height}{Height of saved plot; defaults to 6}

\item{Quality}{Quality of saved plot (dpi); defaults to 900}

\item{File_Type}{File type of saved plot; defaults to "jpg"}

\item{Border_Space_Left}{A continuous value specifying the boundary between the left hand side plot labels and the left margin; defaults to 1}

\item{Border_Space_Right}{A continuous value specifying the boundary between the right hand side plot labels and the right margin; defaults to 1}
}
\value{
Image of Single Forest Plot is saved to the current directory and ggplot object is saved
}
\description{
Title Create forest plots of test statistics of regression model covariates and GWAS summary statistics for SNPs of interest
}
\examples{
Forest_Plot_Model_LM <- Forest_Plot(Data_Sets = c("ModelSumLM", "ModelSumLM"),
             Names = c("BMI LM (1)", "BMI LM (2)"),
             Model_Reference = TRUE,
             Test_Statistic = "BETA", #could be OR
             Display_P_Value_Column = TRUE,
             Display_Test_Stat_Se_Column = TRUE,
             X_Axis_Separation = 0.8,
             Border_Space_Left = 30,
             Border_Space_Right = 75,
             Strips = TRUE,
             Pre_Calculated_CIs = FALSE,
             Legend_Title = "Model",
             Left_Title = "Covariate",
             Test_Stat_Se_Title = "BETA (SE)",
             File_Name = "Forest_Plot", Width =10, Height = 6, Quality = 900,
             File_Type = "jpg"
             )




Forest_Plot_Model_GLM <- Forest_Plot(Data_Sets = c("ModelSumGLM", "ModelSumGLM"),
                                     Names = c("Dementia GLM (1)", "Dementia GLM (2)"),
                                     Model_Reference = TRUE,
                                     Test_Statistic = "OR",
                                     X_Axis_Text_Resolution = 2,
                                     Border_Space_Left = 2.8,
                                     Border_Space_Right = 4.7,
                                     Display_Test_Stat_CI_Column = TRUE,
                                     Display_P_Value_Column = TRUE,
                                     X_Axis_Separation = 0.02,
                                     Pre_Calculated_CIs = FALSE,
                                     Legend_Title = "Model",
                                     Left_Title = "Covariate",
                                     P_Value_Title = "p-value",
                                     Test_Stat_Se_Title = "OR (CI)",
                                     File_Name = "Forest_Plot", Width =10, Height = 6, Quality = 900,
                                     File_Type = "jpg"
)



Forest_Plot_SNPs_BETA <- Forest_Plot(Data_Sets = c("Household_Income_Sum_Stats", "Intelligence_Sum_Stats"),
                                     Names = c("Income", "IQ"),
                                     Model_Reference = FALSE,
                                     Line_Space = 1.5,
                                     Border_Space_Right = 40,
                                     Border_Space_Left = 20,
                                     Test_Statistic = "BETA", #could be OR
                                     Display_Test_Stat_Se_Column = TRUE,
                                     Display_P_Value_Column = TRUE,
                                     X_Axis_Separation = 0.01,
                                     Pre_Calculated_CIs = FALSE,
                                     X_Axis_Text_Resolution = 2,
                                     Legend_Title = "Study",
                                     Left_Title = "SNP",
                                     P_Value_Title = "p-value",
                                     Test_Stat_Se_Title = "BETA (SE)",
                                     Match_Allele_Direction = TRUE,
                                     Match_Allele_Study = "Household_Income_Sum_Stats",
                                     Selected_SNPs = c("rs74832835",  "rs1157671",   "rs1790177",
                                                       "rs9508063",   "rs225682",  "rs56201315" ),
                                     File_Name = "Forest_Plot", Width =10, Height = 6, Quality = 900,
                                     File_Type = "jpg"
)




Forest_Plot_SNPs_BETA_Peak_Finder <- Forest_Plot(Data_Sets = c("Income_Peaks", "Intelligence_Peaks"),
                                                X_Axis_Separation = 0.05,
                                                File_Name = "Forest_Plot", Width =10, Height = 9, Quality = 900,
                                                File_Type = "jpg"
)



Forest_Plot_SNPs_OR <- Forest_Plot(Data_Sets = c("LbDementia_Sum_Stats", "LbDementia_Sum_Stats"),
                                   Names = c("Dementia (1)", "Dementia (2)"),
                                   Model_Reference = FALSE,
                                   Line_Space = 1.1,
                                   Border_Space_Left = 7,
                                   Border_Space_Right = 30,
                                   Test_Statistic = "OR",
                                   Display_P_Value_Column = TRUE,
                                   Display_Test_Stat_CI_Column = TRUE,
                                   X_Axis_Separation = 0.2,
                                   Pre_Calculated_CIs = FALSE,
                                   Legend_Title = "Study",
                                   Left_Title = "SNP",
                                   P_Value_Title = "p-value",
                                   Test_Stat_Se_Title = "OR (CI)",
                                   Match_Allele_Direction = TRUE,
                                   Match_Allele_Study = "LbDementia_Sum_Stats",
                                   Selected_SNPs = c("rs59867714","rs7913723","rs79007041",
                                                     "rs34624328", "rs492457",
                                                     "rs7974838"  ),
                                   File_Name = "Forest_Plot", Width =10, Height = 6, Quality = 900,
                                   File_Type = "jpg"
)




}
