% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intelligence_Sum_Stats.R
\docType{data}
\name{Intelligence_Sum_Stats}
\alias{Intelligence_Sum_Stats}
\title{100,000 SNP Sample Subset of GWAS Summary Statistics of Intelligence}
\format{
A data frame with 100,000 rows (randomly selected) and 14 variables from: https://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/GCST006001-GCST007000/GCST006250/sumstats/README_SavageJansen_2018_intelligence_metaanalysis.txt
}
\usage{
data(Intelligence_Sum_Stats)
}
\description{
A dataset containing a subset of the summary statistics from the Savage Jansen 2018 Intelligence GWAS
The variables are as follows:
}
\details{
\itemize{
\item SNP: rs number
\item UNIQUE_ID: unique SNP id based on chromosome, position
\item CHR: chromosome number
\item POS: base pair position reported on GRCh37
\item A1: effect allele
\item A2: non-effect allele
\item EAF_HRC: effect allele frequency in the Haplotype Reference Consortium reference panel (HRC)
\item Zscore: meta-analysis Z score
\item stdBeta: standardized beta coefficient*
\item SE: standard error of the beta coefficient*
\item P: P-value
\item N_analyzed: sample size
\item minINFO: minimum info score (SNP quality measure) across all cohorts
\item EffectDirection: direction of the effect in each of the cohorts, order: COGENT,GENR,GfG,IMAGEN,NESCOG,BLTS-A,BLTS-C,RS,S4S,STSA-SATSA+GENDER,STSA-HARMONY,DTR-L,DTR-M,STR,TEDS1,TEDS2,UKB-ts,UKB-wb,HIQ

}
}
\keyword{datasets}
