% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Description_METASOFT_Example_Input.R
\docType{data}
\name{METASOFT_Example_Input}
\alias{METASOFT_Example_Input}
\title{100,000 SNP Sample Subset of METASOFT_File_Gen() output from Intelligence_Sum_Stats & Household_Income_Sum_Stats}
\format{
A data frame with 100,000 rows (randomly selected) and 5 variables
}
\usage{
data(METASOFT_Example_Input)
}
\description{
The variables are as follows:
}
\details{
\itemize{
\item ID_DIR: Directionally Ordered RSID
\item Intelligence_Sum_Stats_BETA: BETA value from Intelligence_Sum_Stats SNP
\item Intelligence_Sum_Stats_SE: SE value from Intelligence_Sum_Stats SNP
\item Household_Income_Sum_Stats_BETA: BETA value from Household_Income_Sum_Stats_BETA SNP
\item Household_Income_Sum_Stats_SE: SE value from Household_Income_Sum_Stats_SE SNP

}
}
\keyword{datasets}
