% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METASOFT_File_Gen.R
\name{METASOFT_File_Gen}
\alias{METASOFT_File_Gen}
\title{Title Create datasets of values of interest across multiple GWAS datasets for meta-analysis}
\usage{
METASOFT_File_Gen(
  Data_Sets = c(),
  Chromosome_Columns = c(),
  Reference_Alleles = c(),
  Effect_Alleles = c(),
  Position_Columns = c(),
  SNP_ID_Columns = c(),
  Beta_Columns = c(),
  Standard_Error_Columns = c(),
  PValue_Columns = c(),
  Spec_CHROM = NULL,
  Match_Allele_Direction = TRUE,
  Match_Allele_Study = NULL,
  Output = c("BETA", "SE"),
  File_Name = "METASOFT_Format",
  File_Type = "txt"
)
}
\arguments{
\item{Data_Sets}{These are a list of GWAS summary statistics to be evaluated; defaults to empty list.}

\item{Chromosome_Columns}{These are a list of manual chromosome column names for the data sets used in the order specified; defaults to c()}

\item{Reference_Alleles}{These are a list of manual reference allele column names for the data sets used in the order specified; defaults to c()}

\item{Effect_Alleles}{These are a list of manual effect allele column names for the data sets used in the order specified; defaults to c()}

\item{Position_Columns}{These are a list of manual position column names for the data sets used in the order specified; defaults to c()}

\item{SNP_ID_Columns}{These are a list of manual SNP ID column names for the data sets used in the order specified; defaults to c()}

\item{Beta_Columns}{These are a list of manual BETA column names for the data sets used in the order specified; defaults to  c()}

\item{Standard_Error_Columns}{These are a list of manual SE column names for the data sets used in the order specified; defaults to c()}

\item{PValue_Columns}{These are a list of manual P column names for the data sets used in the order specified; defaults to c()}

\item{Spec_CHROM}{Particular Chromosome to select for a subsetted analysis}

\item{Match_Allele_Direction}{Do you want to match the allele directions to avoid flips (T/F); defaults to T}

\item{Match_Allele_Study}{If the above is T, which data sets should be the reference point for effect allele direction; defaults to "LbDementia_Sum_Stats"; THIS SHOULD BE THE NAME YOU ASSIGN}

\item{Output}{A selection of columns to isolate, in order, alternating by dataset; defaults to c(BETA, SE)}

\item{File_Type}{File type of saved data; defaults to "txt"}

\item{File_NameFile}{name to save data as; defaults to "METASOFT_Format"}
}
\value{
Data frame of information combined across the datasets
}
\description{
Title Create datasets of values of interest across multiple GWAS datasets for meta-analysis
}
\examples{
Combined <- METASOFT_File_Gen(Data_Sets = c("Intelligence_Peaks"))




}
