% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_METASOFT_Run.R
\name{METASOFT_Run}
\alias{METASOFT_Run}
\title{METASOFT Wrapper}
\usage{
METASOFT_Run(..., session = NULL)
}
\arguments{
\item{Data}{These are the harmonised summary statistics (either file path(s), environmental object(s) or a variable(s) containing such contents) to be run through the METASFOT wrapper. These must match the METASFOT input file format and can be generated with METASOFT_File_Gen(); defaults to NULL}

\item{options}{Character vector of extra METASOFT options (e.g. c("-verbose", "mvalue"); defaults to character()/NULL}

\item{REGENIE_Cols}{Include REGENIE format-style columns in output; defaults to TRUE}

\item{Verbose}{Prevent display of progress bar as function is running and instead show key milestone outputs/messages (mainly for debugging purposes); defaults to FALSE.}
}
\value{
METASOFT output file is allocated to specified object
}
\description{
METASOFT Wrapper
}
\examples{
METASOFT_Output <- METASOFT_Run(Data = METASOFT_Example_Input)
}
