% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miami_Plot.R
\name{Miami_Plot}
\alias{Miami_Plot}
\title{Annotated Miami Plot}
\usage{
Miami_Plot(
  Top_Data = Intelligence_Sum_Stats,
  Bottom_Data = Household_Income_Sum_Stats,
  Random_Selection = 10000,
  Draft_Plot = F,
  Top_Chromosome_Column = "CHROM",
  Top_Position_Column = "GENPOS",
  Top_SNP_ID_Column = "ID",
  Top_PValue_Column = "P",
  Top_Colour_One = "green",
  Top_Colour_Two = "purple",
  Bottom_Colour_One = "green",
  Bottom_Colour_Two = "purple",
  Bottom_Chromosome_Column = "CHROM",
  Bottom_Position_Column = "GENPOS",
  Bottom_SNP_ID_Column = "ID",
  Bottom_PValue_Column = "P",
  Top_Title = "Plot1",
  Bottom_Title = "Plot2",
  Bottom_Title_Size = 35,
  Top_Title_Size = 35,
  Y_Axis_Text_Size = 30,
  Y_Axis_Title_Size = 35,
  Chromosome_Label_Size = 30,
  Top_Sig_Line_Colour = "red",
  Bottom_Sig_Line_Colour = "red",
  Top_Sig_Line_Type = "dashed",
  Bottom_Sig_Line_Type = "dashed",
  Top_Sig_Threshold = 5e-08,
  Bottom_Sig_Threshold = 5e-08,
  Top_Sig_Line_Width = 0.5,
  Bottom_Sig_Line_Width = 0.5,
  Point_Size = 2.5,
  Top_Label_Index = T,
  Bottom_Label_Index = T,
  Label_Size = 6,
  Label_Angle = 30,
  Top_Label_Colour = "black",
  Bottom_Label_Colour = "black",
  Top_Colour_Of_Index = "red",
  Top_Chromosome_Labels = c(2, 6, 12, 14, 16, 20),
  Top_Chromosome_Index = c(2, 6, 12, 14, 16, 20),
  Bottom_Chromosome_Labels = c(2, 6, 12, 14, 16, 20),
  Bottom_Chromosome_Index = c(2, 6, 12, 14, 16, 20),
  Bottom_Colour_Of_Index = "red",
  Top_Colour_Index = T,
  Bottom_Colour_Index = T,
  Top_Condense_Scale = T,
  Bottom_Condense_Scale = T,
  Top_Break_Point = 1e-10,
  Bottom_Break_Point = 1e-10,
  File_Name = "Miami_Plot",
  Width = 30,
  Height = 15,
  Quality = 900,
  File_Type = "jpg"
)
}
\arguments{
\item{Top_Data}{These are the summary statistics to be plotted on the top panel; defaults to Intelligence_Sum_Stats}

\item{Bottom_Data}{These are the summary statistics to be plotted on the bottom panel; defaults to Household_Income_Sum_Stats}

\item{Random_Selection}{Manually select a random subset of SNPs to plot, given that Draft_Plot is TRUE; defaults to 10000}

\item{Draft_Plot}{Specify whether a subset of SNPs, specified by Random_Selection is plotted; defaults to F}

\item{Top_Chromosome_Column}{Manually specify chromosome column for top panel; defaults to "CHROM"}

\item{Top_Position_Column}{Manually specify chromosome position column for top panel; defaults to "GENPOS"}

\item{Top_SNP_ID_Column}{Manually specify SNP ID column for top panel; defaults to "ID"}

\item{Top_PValue_Column}{Manually specify P Value column for top panel; defaults to "P"}

\item{Top_Colour_One}{The first colour (odd chromosome numbers) to use on the top plot; defaults to "green"}

\item{Top_Colour_Two}{The second colour (even chromosome numbers) to use on the top plot; defaults to "purple"}

\item{Bottom_Colour_One}{The first colour (odd chromosome numbers) to use on the bottom plot; defaults to "green"}

\item{Bottom_Colour_Two}{The second colour (even chromosome numbers) to use on the bottom plot; defaults to "purple"}

\item{Bottom_Chromosome_Column}{Manually specify chromosome column for bottom panel; defaults to "CHROM"}

\item{Bottom_Position_Column}{Manually specify chromosome position column for bottom panel; defaults to "GENPOS"}

\item{Bottom_SNP_ID_Column}{Manually specify SNP ID column for bottom panel; defaults to "ID"}

\item{Bottom_PValue_Column}{Manually specify P Value column for bottom panel; defaults to "P"}

\item{Top_Title}{Title to be displayed on the top panel; defaults to "Plot1"}

\item{Bottom_Title}{Title to be displayed on the bottom panel; defaults to "Plot2"}

\item{Bottom_Title_Size}{Size of bottom panel title; defaults to 35}

\item{Top_Title_Size}{Size of top panel title; defaults to 35}

\item{Y_Axis_Text_Size}{Size of Y axes text; defaults to 30}

\item{Y_Axis_Title_Size}{Size of Y axes titles; defaults to 35}

\item{Chromosome_Label_Size}{Size of chromosome numbers; defaults to 30}

\item{Top_Sig_Line_Colour}{Colour of significance threshold on top panel; defaults to "red"}

\item{Bottom_Sig_Line_Colour}{Colour of significance threshold on bottom panel; defaults to "red"}

\item{Top_Sig_Line_Type}{Type of significance threshold line on top panel; defaults to "dashed"}

\item{Bottom_Sig_Line_Type}{Type of significance threshold line on bottom panel; defaults to "dashed"}

\item{Top_Sig_Threshold}{Value of significance threshold on top panel; defaults to 5e-8}

\item{Bottom_Sig_Threshold}{Value of significance threshold on bottom panel; defaults to 5e-8}

\item{Top_Sig_Line_Width}{Width of significance threshold line on top panel; defaults to 0.5}

\item{Bottom_Sig_Line_Width}{Width of significance threshold line on bottom panel; defaults to 0.5}

\item{Point_Size}{Size of points; defaults to 2.5}

\item{Top_Label_Index}{Label the index SNPs with ID provided on top panel; defaults to T}

\item{Bottom_Label_Index}{Label the index SNPs with ID provided on bottom panel; defaults to T}

\item{Label_Size}{Size of labels used; defaults to 6}

\item{Label_Angle}{Angle of labels used; defaults to 30}

\item{Top_Label_Colour}{Colour of index SNP label on top panel; defaults to "black"}

\item{Bottom_Label_Colour}{Colour of index SNP label on bottom panel; defaults to "black"}

\item{Top_Colour_Of_Index}{Colour of index SNP highlight on top panel; defaults to "red"}

\item{Top_Chromosome_Labels}{Manually specify which chromosome index SNPs should be labelled on the top plot; defaults to  c(2,6,12,14,16,20)}

\item{Top_Chromosome_Index}{Manually specify which chromosome index SNPs should be circled/highlighted on the top plot; defaults to  c(2,6,12,14,16,20)}

\item{Bottom_Chromosome_Labels}{Manually specify which chromosome index SNPs should be labelled on the bottom plot; defaults to  c(2,6,12,14,16,20)}

\item{Bottom_Chromosome_Index}{Manually specify which chromosome index SNPs should be circled/highlighted on the bottom plot; defaults to  c(2,6,12,14,16,20)}

\item{Bottom_Colour_Of_Index}{Colour of index SNP highlight on bottom panel; defaults to "red"}

\item{Top_Colour_Index}{Highlight the index SNP on top panel; defaults to T}

\item{Bottom_Colour_Index}{Highlight the index SNP on bottom panel; defaults to T}

\item{Top_Condense_Scale}{Condense scale on top panel; defaults to T}

\item{Bottom_Condense_Scale}{Condense scale on bottom panel; defaults to T}

\item{Top_Break_Point}{Value where y axis condenses on top panel; defaults to "red"}

\item{Bottom_Break_Point}{Value where y axis condenses on bottom panel; defaults to "red"}

\item{File_Name}{File name to save plot as; defaults to "Miami_Plot"}

\item{Width}{Width of saved plot; defaults to 30}

\item{Height}{Height of saved plot; defaults to 15}

\item{Quality}{Quality of saved plot (dpi); defaults to 900}

\item{File_Type}{File type of saved plot; defaults to "jpg"}
}
\value{
Image of Miami Plot is saved to the current directory and ggplot object is saved for top and bottom plots in addition to a combined grob
}
\description{
Annotated Miami Plot
}
\examples{
Miami_Plot <- Miami_Plot(Top_Data = Intelligence_Sum_Stats, Bottom_Data = Household_Income_Sum_Stats,
                        Top_Chromosome_Column = "CHROM",
                        Top_Position_Column = "GENPOS", Top_SNP_ID_Column = "ID",
                        Top_PValue_Column = "P",
                        Top_Colour_One = "green", Top_Colour_Two = "purple",
                        Bottom_Colour_One = "green", Bottom_Colour_Two = "purple",
                        Bottom_Chromosome_Column = "CHROM",
                        Bottom_Position_Column = "GENPOS", Bottom_SNP_ID_Column = "ID",
                        Bottom_PValue_Column = "P",
                        Top_Title = "Intelligence", Bottom_Title = "Household Income",
                        File_Name = "Miami_Plot")


}
