% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{data}
\name{Model}
\alias{Model}
\title{Logistic regression Model Object of Dementia cases derived from simulated data}
\format{
A logistic regression base R model object
}
\usage{
data(Model)
}
\description{
The object was created as follows: Model <- glm(Dementia ~ Ethnicity + Age + Sex + BMI + Location, data = Fake_Demo_Data)
}
\keyword{datasets}
