% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model2.R
\docType{data}
\name{Model2}
\alias{Model2}
\title{Linear regression Model Object of BMI values derived from simulated data}
\format{
A liner regression base R model object
}
\usage{
data(Model2)
}
\description{
The object was created as follows: Model2 <- lm(BMI ~ Ethnicity + Age + Sex + Dementia + Location, data = Fake_Demo_Data)
}
\keyword{datasets}
