% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_Model_Munge.R
\name{Model_Munge}
\alias{Model_Munge}
\title{Munged, Covariate & Modelling Data}
\usage{
Model_Munge(..., session = NULL)
}
\arguments{
\item{...}{Shadow argument to detect arguments passed; defaults to empty list()}

\item{Model_Object}{Base R model generated from running an lm() or glm(); defaults to NULL}

\item{Verbose}{Prevent display of progress bar as function is running and instead show key milestone outputs/messages (mainly for debugging purposes); defaults to FALSE}
}
\value{
Munged dataframe amenable to plotting test statistics which can then be passed to other MiamiR functions
}
\description{
Munged, Covariate & Modelling Data
}
\examples{
 Model_Munged <- Model_Munge(Model_Object = "Model_One")
}
