% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Description_Model_One.R
\docType{data}
\name{Model_One}
\alias{Model_One}
\title{Model object derived from Fake_Demo_Data}
\format{
A model object
}
\usage{
data(Model_One)
}
\description{
A model object containing coefficients for passed variables
}
\details{
(BMI ~ Age + Sex + Ethnicity + Location, data = Fake_Demo_Data)

The variables are as follows:

\itemize{
\item Age = Age (in years)
\item Sex = Genetic Sex
\item BMI = Continuous body mass index
\item Ethnicity = Broad self-report ethnicity
\item Location = Sample site, grouped by dwelling type

}
}
\keyword{datasets}
