% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSumGLM.R
\docType{data}
\name{ModelSumGLM}
\alias{ModelSumGLM}
\title{Data frame of covariate effects from glm object "Model", created using the Model_Munge() function.}
\format{
A data frame with 9 rows and 6 variables (simulated)
}
\usage{
data(ModelSumGLM)
}
\description{
A dataset containing processed test statistics from the dementia glm model
The variables are as follows:
}
\details{
\itemize{
\item Covariate = Name of covariate which test statistics represent
\item OR = Odds Ratio
\item SE = Standard Error of the effect estimate/OR
\item P = p-value
\item group = The overall group which covariates, particularly factor covarites fall into
\item Reference = The reference level for factor covariates, derived from meta-data

}
}
\keyword{datasets}
