% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_Finder.R
\name{Peak_Finder}
\alias{Peak_Finder}
\title{Title}
\usage{
Peak_Finder(
  Data_Sets = c("HillWD_31844048_household_Income.txt",
    "SavageJansen_2018_intelligence_metaanalysis"),
  Names = NULL,
  Chromosome_Columns = c(),
  Reference_Alleles = c(),
  Effect_Alleles = c(),
  Position_Columns = c(),
  SNP_ID_Columns = c(),
  Beta_Columns = c(),
  Standard_Error_Columns = c(),
  PValue_Columns = c(),
  Spec_CHROM = NULL,
  Peak_Separation = 1e+06,
  File_Name = "Peaks",
  File_Type = "txt"
)
}
\arguments{
\item{Data_Sets}{These are a list of GWAS summary statistics to be evaluated; defaults to c("LbDementia_Sum_Stats", "Intelligence_Sum_Stats")}

\item{Names}{These are a list of names for the data sets to be labelled with in the output file - this also specifies the order down the page; defaults to c("Dementia", "Intelligence")}

\item{Chromosome_Columns}{These are a list of manual chromosome column names for the data sets used in the order specified; defaults to c()}

\item{Reference_Alleles}{These are a list of manual reference allele column names for the data sets used in the order specified; defaults to c()}

\item{Effect_Alleles}{These are a list of manual effect allele column names for the data sets used in the order specified; defaults to c()}

\item{Position_Columns}{These are a list of manual position column names for the data sets used in the order specified; defaults to c()}

\item{SNP_ID_Columns}{These are a list of manual SNP ID column names for the data sets used in the order specified; defaults to c()}

\item{Beta_Columns}{These are a list of manual BETA column names for the data sets used in the order specified; defaults to  c()}

\item{Standard_Error_Columns}{These are a list of manual SE column names for the data sets used in the order specified; defaults to c()}

\item{PValue_Columns}{These are a list of manual P column names for the data sets used in the order specified; defaults to c()}

\item{Peak_Separation}{Distnance in BP necessary from a lead SNP for another region of association to begin; defaults to 100K}

\item{File_Name}{File name to save data as; defaults to "Peaks"}

\item{File_Type}{File type of saved data; defaults to "txt"}
}
\value{
SNPs split by study saved
}
\description{
Title
}
\examples{


Peak_Finder_Outcome <- Peak_Finder(Data_Sets = Locations,
                                Peak_Separation = 2000000,
                                File_Name = "Peaks",
                                Spec_CHROM = 1,
                                File_Type = "txt")




}
