% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_Regional_Plot.R
\name{Regional_Plot}
\alias{Regional_Plot}
\title{Annotated, Customised Regional Plots}
\usage{
Regional_Plot(..., session = NULL)
}
\arguments{
\item{Data}{These are the summary statistics (either file path(s), environmental object(s) or a variable(s) containing such contents) to be plotted; defaults to NULL}

\item{Chromosome_Column}{Manually specify chromosome column; defaults to NULL, leading to auto-detection}

\item{Position_Column}{Manually specify chromosome genomic position column; defaults to NULL, leading to auto-detection}

\item{SNP_ID_Column}{Manually specify SNP ID column; defaults to NULL, leading to auto-detection}

\item{PValue_Column}{Manually specify P Value column; defaults to NULL, leading to auto-detection}

\item{Chromosomes}{Manually specify which chromosomes from the summary statistics to run the function on; defaults to NULL, leading to all being included/absent filtering}

\item{MIN_GENE_GAP}{Minimum genomic distance between adjacent annotation anchoring/positioning for Gene_Tracks body names, when Gene_Tracks = TRUE; defaults to NULL and if not passed is calculated from 0.3 X the Region_Window size specified.}

\item{Separation_Distance}{The minimum genomic distance between significant regions of interest/lead SNPs to be defined as distinct peaks, defined by P_threshold value; defaults to 1e6}

\item{Region_Window}{Distance to be included up and downstream of the lead SNP in a region of interest; defaults to 5e5}

\item{Gene_Label_Size}{Size of the text allocated to the bodies displayed in the gene panel, when Gene_Tracks is TRUE; defaults to 10}

\item{Gene_Structure_Size_Exons}{Size of the exon object allocated to the bodies displayed in the gene panel, when Gene_Tracks is TRUE; defaults to 12}

\item{Gene_Structure_Size_Introns}{Size of the intron object allocated to the bodies displayed in the gene panel, when Gene_Tracks is TRUE; defaults to 2}

\item{Gene_Structure_Colour_Exons}{Colour of the exon object allocated to the bodies displayed in the gene panel, when Gene_Tracks is TRUE; defaults to black}

\item{Gene_Structure_Colour_Introns}{Colour of the intron object allocated to the bodies displayed in the gene panel, when Gene_Tracks is TRUE; defaults to black}

\item{Gene_Panel_Background_Colour}{Colour of the background panel surrounding the bodies displayed in the gene panel, when Gene_Tracks is TRUE; defaults to azure}

\item{Sense_Arrow}{Show direction of transcription arrow in the gene panel, when Gene_Tracks is TRUE; defaults to TRUE}

\item{Sense_Arrow_Colour}{Colour of direction of transcription arrow in the gene panel, when Gene_Tracks is TRUE; defaults to black}

\item{Sense_Arrow_Head_Size}{Size of direction of transcription arrow head/tip in the gene panel, when Gene_Tracks is TRUE; defaults to 7}

\item{Sense_Arrow_Body_Thickness}{Width of direction of transcription arrow body in the gene panel, when Gene_Tracks is TRUE; defaults to NULL and when not passed is calculated based on Gene_Structure_Size_Introns X 0.6}

\item{Sense_Arrow_Body_Length}{Length of direction of transcription arrow body in the gene panel, when Gene_Tracks is TRUE; defaults to 0.02}

\item{Sense_Arrow_Gene_Gap}{Minimum genomic distance (spacing) between end of Gene_Tracks bodies and beginning of direction of transcription arrow rendering, when Gene_Tracks = TRUE; defaults to 0.005 (factor of view window, proxy for Region_Window)}

\item{Recombination_Line}{Display recombination rate via line plot overlay across region shown; defaults to TRUE}

\item{Recombination_Line_Colour}{Colour of recombination rate line across region shown when Recombination Line is TRUE; defaults to darkred}

\item{Recombination_Axis_Title}{Y Axis Title of recombination rate scale when Recombination_Line is TRUE; defaults to Recombination Rate (cM/Mb)}

\item{Recombination_Axis_Text_Size}{Y Axis text size of recombination rate scale when Recombination_Line is TRUE; defaults to NULL and if not passed is based on inherited Single_Plot() Y_Axis_Text_Size and its default value}

\item{Recombination_Axis_Title_Size}{Y Axis title size of recombination rate scale when Recombination_Line is TRUE; defaults to NULL and if not passed is based on inherited Single_Plot() Y_Axis_Title_Size and its default value}

\item{Recombination_Line_Thickness}{Width of recombination rate line across region shown when Recombination_Line is TRUE; defaults to 1}

\item{Recombination_Line_Type}{Line style of recombination rate line across region shown when Recombination_Line is TRUE; defaults to dotted}

\item{P_threshold}{Maximum value of P (minimum LOG10P) allowed for a SNP to be classified as initiating detection as the lead SNP of a peak in the function; defaults to 5e-8}

\item{Point_Colour}{Colour of data points shown; defaults to darkblue}

\item{LD_Legend_Title_Size}{Size of title in Auto_LD Legend when both LD_Legend_On and Auto_LD are TRUE; defaults to 30}

\item{LD_Legend_Text_Size}{Size of text in Auto_LD Legend key when both LD_Legend_On and Auto_LD are TRUE; defaults to 20}

\item{LD_Legend_Location}{Relative location of LD Legend in the top panel when both LD_Legend_On and Auto_LD are TRUE; defaults to "Top Right"}

\item{LD_Legend_On}{Show Auto_LD Legend, when Auto_LD is TRUE; defaults to TRUE}

\item{Gene_Legend_Location}{Relative location of Gene_Tracks Legend in the bottom panel when both Gene_Legend_On and Gene_Tracks are TRUE; defaults to "Right Outside"}

\item{Gene_Legend_On}{Show Gene_Legend, when Gene_Tracks is TRUE; defaults to TRUE}

\item{Gene_Legend_Title_Size}{Size of title in the Gene_Tracks legend when both Gene_Legend_On and Gene_Tracks are TRUE; defaults to 30}

\item{Gene_Legend_Text_Size}{Size of text in the Gene_Tracks Legend key when both Gene_Legend_On and Gene_Tracks are TRUE; defaults to 20}

\item{Gene_Legend_Title}{Gene_Tracks Legend heading when both Gene_Legend_On and Gene_Tracks are TRUE; defaults to Gene Biotype}

\item{LD_Legend_Title}{Auto_LD Legend heading when both LD_Legend_On and Auto_LD are TRUE; defaults to "R\u00b2 LD" - R² LD}

\item{Genome_Build}{Reference genome to base required plot panels and annotations around (e.g. Auto_LD and Gene_Tracks functionality); defaults to grch38}

\item{Population}{Reference population for Auto_LD query; defaults to "EUR"}

\item{Gene_Biotype}{Filter bodies shown when Gene_Tracks is TRUE by biotype, e.g. coding; defaults to NULL allowing for all available biotypes to be shown.}

\item{Gene_Biotype_Colour}{Colour of bodies shown when Gene_Tracks is TRUE, either overall or stratified by biotype, e.g. c("blue", "red); defaults to NULL allowing for default colouring panel to be allotted}

\item{Gene_Name}{Filter bodies shown when Gene_Tracks is TRUE by body name, e.g. "DMD"; defaults to NULL allowing for all available bodies to be shown.}

\item{Auto_LD_Region_Size}{Region to query centered around the ascertained lead SNP in the peak region, when Auto_LD is TRUE; defaults to 1e6}

\item{Auto_LD}{Query for LD via the included LDlinkR::LDproxy functionality, based around Genome_Build & Population arguments, centered around the ascertained lead SNP in the region shown; defaults to FALSE}

\item{Reference_Allele_Column}{Manually specify reference allele column; defaults to NULL, leading to auto-detection}

\item{Effect_Allele_Column}{Manually specify effect allele column; defaults to NULL, leading to auto-detection}

\item{Verbose}{Prevent display of progress bar as function is running and instead show key milestone outputs/messages (mainly for debugging purposes); defaults to FALSE.}

\item{Interactive}{View the output as amenable to interactive plotly object, which can be used to inspect data, with key overlays (more for shiny App usage); defaults to FALSE}

\item{Gene_Tracks}{Display gene panel below plot with specified bodies within the genomic region shown; defaults to TRUE}

\item{.dots}{Shadow argument to detect inherited arguments passed and modified from the Single_Plot() function; defaults to list()}

\item{Manual_Decimal_Places}{Rounding of places to target regarding the X axis and genomic positions (Mb); defaults to NULL, and is automatically assigned unless specified}

\item{P_Value_Filter}{Threshold below which to ignore, primarily non-important points; defaults to NULL}

\item{Gene_Label_Buffer}{Units of space to add/jitter between each gene label near the edge of the Gene_Tracks panel to align inwards; defaults to 1}

\item{Gene_Label_Vertical_Spacing}{Units of space to add between each gene label and gene track/body in the Gene_Tracks panel; defaults to 0.375}

\item{Panel_Ratio}{Specify ratio when of Plot_Panel area to Annotation area, Gene_Tracks is TRUE; defaults to NULL, leading to auto calculation unless manually specified}

\item{Plot_Panel}{Decide whether to include the Plot area above the gene annotation; defaults to TRUE}

\item{Target_Position_Breaks}{Ideal number of positions to annotate on the genomic position/X axis; defaults to 6}

\item{Gene_Track_Vertical_Spacing}{Units of space to add between each gene track in the Gene_Tracks panel; defaults to 0.75}

\item{Recombination_Axis_Title_Vjust}{Number of units of blank space between Recombination_Axis_Title element area and adjacent axis; defaults to 7}

\item{Manual_LD}{Specify manual LD matrices, in PLINK format to be projected on to the points via colour binning; defaults to NULL}
}
\value{
Image of Regional Plot(s) is allocated to specified object and the resulting ggplot object can then be saved to an image
}
\description{
Annotated, Customised Regional Plots
}
\examples{
 Regional_Plot <- Regional_Plot(Data = Household_Income_Sum_Stats, Chromosomes = 1)
}
