% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_Segregate_HLA_Plot.R
\name{Segregate_HLA_Plot}
\alias{Segregate_HLA_Plot}
\title{Annotated, Customised HLA-WAS Plots}
\usage{
Segregate_HLA_Plot(
  Data = NULL,
  Verbose = FALSE,
  Title = NULL,
  Title_On = TRUE,
  Title_Size = 20,
  X_Axis_Title = "HLA Genes",
  Y_Axis_Title = "-log₁₀(P)",
  X_Axis_Title_Size = 20,
  Y_Axis_Title_Size = 20,
  X_Axis_Title_Vjust = -6,
  Y_Axis_Title_Vjust = 6,
  X_Axis_Text_Vjust = -0.2,
  Y_Axis_Text_Vjust = 0.2,
  X_Axis_Text_Size = 15,
  Y_Axis_Text_Size = 15,
  Label_Index = TRUE,
  Label_Size = 4,
  Label_Height = 2,
  Segment_Width = 0.18,
  Segment_Thickness = 0.9,
  Sig_Line_Colour = "red",
  Sig_Line_Type = "dashed",
  Sig_Threshold = 5e-08,
  Sig_Line_Width = 0.6,
  session = NULL
)
}
\arguments{
\item{Data}{These are the summary statistics (either file path(s), environmental object(s) or a variable(s) containing such contents) to be plotted; defaults to NULL}

\item{Verbose}{Prevent display of progress bar as function is running and instead show key milestone outputs/messages (mainly for debugging purposes); defaults to FALSE}

\item{Title}{Manually specify title to be displayed; defaults to NULL, leading to auto-detection of dataframe object/file name}

\item{Title_On}{Toggle on/off whether Title (either manually or automatically determined) displays on final plot; defaults to TRUE}

\item{Title_Size}{Size of title; defaults to 20}

\item{X_Axis_Title}{Manually specify X axis title; defaults to "HLA Genes"}

\item{Y_Axis_Title}{Manually specify Y axis title; defaults to "-log\u2081\u2080(P)" - shows as "-log₁₀(P)"}

\item{X_Axis_Title_Size}{Size of X axis title; defaults to 20}

\item{Y_Axis_Title_Size}{Size of Y axis title; defaults to 20}

\item{X_Axis_Title_Vjust}{Number of units of blank space between X_Axis_Title element area and above axis; defaults to -6}

\item{Y_Axis_Title_Vjust}{Number of units of blank space between Y_Axis_Title element area and adjacent axis; defaults to 6}

\item{X_Axis_Text_Vjust}{Number of units of blank space between X_Axis_Text element area and above axis; defaults to -0.2}

\item{Y_Axis_Text_Vjust}{Number of units of blank space between Y_Axis_Text element area and adjacent axis; defaults to 0.2}

\item{X_Axis_Text_Size}{Size of X axis text; defaults to 15}

\item{Y_Axis_Text_Size}{Size of Y axis text; defaults to 15}

\item{Label_Index}{Annotate the index SNPs with ID provided; defaults to TRUE}

\item{Label_Size}{Size of index labels if Label_Index is TRUE; defaults to 4}

\item{Label_Height}{Spacing between the point and the label position; defaults to 2}

\item{Segment_Width}{Half-width of each per-point horizontal segment (total length = 2 * Segment_Width); defaults to 0.18}

\item{Segment_Thickness}{Thickness/line width for the per-point horizontal segments; defaults to 0.9}

\item{Sig_Line_Colour}{Colour of significance threshold line; defaults to "red"}

\item{Sig_Line_Type}{Type of significance threshold line; defaults to "dashed"}

\item{Sig_Threshold}{Value of significance threshold; defaults to standard Bonferroni correction of 5e-8}

\item{Sig_Line_Width}{Width of significance threshold line; defaults to 0.6}
}
\value{
Image of HLA type-specific Line Plot(s) is allocated to specified object and the resulting ggplots object can then be saved to images
}
\description{
Annotated, Customised HLA-WAS Plots
}
\examples{
HLA_Plots <- Segregate_HLA_Plot(Data = "Fake_HLA_100K")

}
