% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Plot.R
\name{Single_Plot}
\alias{Single_Plot}
\title{Single, Annotated Manhattan Plot}
\usage{
Single_Plot(
  Data = Intelligence_Sum_Stats,
  Chromosome_Column = "CHROM",
  Random_Selection = 10000,
  Draft_Plot = F,
  Position_Column = "GENPOS",
  SNP_ID_Column = "ID",
  PValue_Column = "P",
  Chromosome_Labels = c(2, 6, 12, 14, 16, 20),
  Chromosome_Index = c(2, 6, 12, 14, 16, 20),
  Title = "Plot1",
  Title_Size = 35,
  Y_Axis_Text_Size = 30,
  Y_Axis_Title_Size = 35,
  X_Axis_Title_Size = 35,
  X_Axis_Title = "Chromosome",
  Chromosome_Label_Size = 30,
  Chromosome_Colours = c("blue", "turquoise"),
  Sig_Line_Colour = "red",
  Sig_Line_Type = "dashed",
  Sig_Threshold = 5e-08,
  Sig_Line_Width = 0.5,
  Point_Size = 2.5,
  Label_Index = T,
  Label_Size = 6,
  Label_Angle = 30,
  Label_Colour = "black",
  Colour_Of_Index = "red",
  Colour_Index = T,
  Condense_Scale = T,
  Break_Point = 1e-10,
  File_Name = "Manhattan_Plot",
  Width = 30,
  Height = 15,
  Quality = 900,
  File_Type = "jpg"
)
}
\arguments{
\item{Data}{These are the summary statistics to be plotted; defaults to Intelligence_Sum_Stats}

\item{Chromosome_Column}{Manually specify chromosome column; defaults to "CHROM"}

\item{Random_Selection}{Manually select a random subset of SNPs to plot, given that Draft_Plot is TRUE; defaults to 10000}

\item{Draft_Plot}{Specify whether a subset of SNPs, specified by Random_Selection is plotted; defaults to F}

\item{Position_Column}{Manually specify chromosome position column; defaults to "GENPOS"}

\item{SNP_ID_Column}{Manually specify SNP ID column; defaults to "ID"}

\item{PValue_Column}{Manually specify P Value column; defaults to "P"}

\item{Chromosome_Labels}{Manually specify which chromosome index SNPs should be labelled on the plot; defaults to  c(2,6,12,14,16,20)}

\item{Chromosome_Index}{Manually specify which chromosome index SNPs should be circled/highlighted on the plot; defaults to  c(2,6,12,14,16,20)}

\item{Title}{Title to be displayed; defaults to "Plot1"}

\item{Title_Size}{Size of title; defaults to 35}

\item{Y_Axis_Text_Size}{Size of Y axis text; defaults to 30}

\item{Y_Axis_Title_Size}{Size of Y axis title; defaults to 35}

\item{X_Axis_Title_Size}{Size of X axis title; defaults to 35}

\item{X_Axis_Title}{X axis title; defaults to "Chromosome"}

\item{Chromosome_Label_Size}{Size of chromosome numbers; defaults to 30}

\item{Chromosome_Colours}{Alternating list of two colours for chromosome points (odd/even); defaults to c("blue", "turquoise")}

\item{Sig_Line_Colour}{Colour of significance threshold; defaults to "red"}

\item{Sig_Line_Type}{Type of significance threshold line; defaults to "dashed"}

\item{Sig_Threshold}{Value of significance threshold; defaults to 5e-8}

\item{Sig_Line_Width}{Width of significance threshold line; defaults to 0.5}

\item{Point_Size}{Size of points; defaults to 2.5}

\item{Label_Index}{Label the index SNPs with ID provided; defaults to T}

\item{Label_Size}{Size of labels used; defaults to 6}

\item{Label_Angle}{Angle of labels used; defaults to 30}

\item{Label_Colour}{Colour of index SNP label; defaults to "black"}

\item{Colour_Of_Index}{Colour of index SNP highlight; defaults to "red"}

\item{Colour_Index}{Highlight the index SNP; defaults to T}

\item{Condense_Scale}{Condense scale; defaults to T}

\item{Break_Point}{Value where y axis condenses; defaults to 1e-10}

\item{File_Name}{File name to save plot as; defaults to "Manhattan_Plot"}

\item{Width}{Width of saved plot; defaults to 30}

\item{Height}{Height of saved plot; defaults to 15}

\item{Quality}{Quality of saved plot (dpi); defaults to 900}

\item{File_Type}{File type of saved plot; defaults to "jpg"}
}
\value{
Image of Single Manhattan Plot is saved to the current directory and ggplot object is saved
}
\description{
Single, Annotated Manhattan Plot
}
\examples{
 Manhattan_Plot <- Single_Plot(Data = Intelligence_Sum_Stats,
                             Chromosome_Column = "CHROM",
                             Position_Column = "GENPOS", SNP_ID_Column = "ID",
                             PValue_Column = "P",
                             Title = "Intelligence",
                             Chromosome_Colours = c("blue", "turquoise"),
                             File_Name = "Manhattan_Plot")

}
