% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_Single_With_Regional_Plot.R
\name{Single_With_Regional_Plot}
\alias{Single_With_Regional_Plot}
\title{Customised Manhattan Plot with Inset Regional Plots}
\usage{
Single_With_Regional_Plot(..., Verbose = FALSE, session = NULL)
}
\arguments{
\item{Verbose}{Prevent display of progress bar as function is running and instead show key milestone outputs/messages (mainly for debugging purposes); defaults to FALSE}

\item{Single_Plot}{Object returned from Single_Plot(); defaults to NULL}

\item{Regional_Plot}{Object(s) returned from Regional_Plot(); defaults to NULL}

\item{LD_Legend_Plot}{Specific Regional_Plot() object to derive LD_Legend from; defaults to NULL, leading to first object being utilised}

\item{Gene_Legend_Plot}{Specific Regional_Plot() object to derive Gene_Legend from; defaults to NULL, leading to merged legend being created across and from any objects supplied}

\item{LD_Legend_Scale}{Proportion of default sizing to use when displaying LD key on Single_Plot() background derived from inset Regional_Plot(); defaults to 1}

\item{Gene_Legend_Scale}{Proportion of default sizing to use when displaying Gene Biotype key on Single_Plot() background derived from inset Regional_Plot(); defaults to 1}

\item{LD_Legend_Height}{Proportion of default sizing height to use when anchoring LD key on Single_Plot() background; defaults to NULL}

\item{Gene_Legend_Height}{Proportion of default sizing height to use when anchoring Gene Biotype key on Single_Plot() background; defaults to NULL, leading to max anchoring relative to default or specified Gene_Legend_Preset}

\item{LD_Legend_Preset}{Automatically set anchoring of LD key on Single_Plot() background relative to common positioning; defaults to NULL, leading to "TopRight" being applied}

\item{Gene_Legend_Preset}{Automatically set anchoring of Gene Biotype key on Single_Plot() background relative to common positioning; defaults to NULL, leading to "TopRight" being applied}

\item{Target_Width}{Target save width in inches which the user intends on applying when using ggsave on the combined object; defaults to 30}

\item{Target_Height}{Target save height in inches which the user intends on applying when using ggsave on the combined object; defaults to 15}

\item{Inset_Scale}{Proportion of the full default size of the Regional_Plot() panels to use when overlaying as insets; defaults to 0.17}

\item{Inset_Chromosomes}{Allocated chromosomes where inset based on lead signal/set of SNPs is to be inserted; defaults to NULL leading to all Regional_Plot() objects provided to Regional_Plot argument being inserted in ascending order of significant signal on combined plot and by current order of regional objects}

\item{Preset_Tweak_X}{Add further spacing along the X-axis to the predefined Inset_Preset direction of Regional_Plot() anchoring around signal of interest; defaults to 0}

\item{Preset_Tweak_Y}{Add further spacing along the Y-axis to the predefined Inset_Preset direction of Regional_Plot() anchoring around signal of interest; defaults to 0}

\item{Preset_Regional_Height_Tweak}{Proportion of the full default Inset_Scale height of the Regional_Plot() panels to use when overlaying as insets allowing for further control in 2-Dimensions; defaults to 0.3}

\item{Inset_Preset}{Relative predefined position of Regional_Plot() inset relative to peak of interest; defaults to "Right"}

\item{Chromosome_Column}{Manually specify chromosome column; defaults to NULL, leading to auto-detection}

\item{Peak_Highlight_On}{Draw box around significant signals of interest from which inset expands from; defaults to TRUE}

\item{Peak_Highlight_Colour}{Colour of box around significant signals of interest from which inset expands from where Peak_Highlight_On is TRUE; defaults to "grey40"}

\item{Peak_Highlight_Linetype}{Line type of box around significant signals of interest from which inset expands from where Peak_Highlight_On is TRUE; defaults to "solid"}

\item{Peak_Highlight_Thickness}{Thickness of box around Regional_Plot() inset; defaults to 0.45}

\item{Regional_Inset_Highlight_On}{Draw box around Regional_Plot() inset; defaults to TRUE}

\item{Regional_Inset_Highlight_Colour}{Colour of box around Regional_Plot() inset; defaults to "grey40"}

\item{Regional_Inset_Highlight_Linetype}{Line type of box around Regional_Plot() inset; defaults to "solid"}

\item{Regional_Inset_Highlight_Thickness}{Line thickness of box around Regional_Plot() inset; defaults to 0.45}

\item{Inset_Lines_On}{Draw lines from box around significant signals of interest from which inset expands from to the box around the Regional_Plot() inset; defaults to TRUE}

\item{Inset_Lines_Colour}{Colour of lines from box around significant signals of interest from which inset expands from to the box around the Regional_Plot() inset; defaults to "black"}

\item{Inset_Lines_Linetype}{Type of lines from box around significant signals of interest from which inset expands from to the box around the Regional_Plot() inset; defaults to "dotted"}

\item{Inset_Lines_Thickness}{Thickness of lines from box around significant signals of interest from which inset expands from to the box around the Regional_Plot() inset; defaults to 0.45}

\item{Inset_Lines_Gap_Shading_On}{Shade space between lines from box around significant signals of interest from which inset expands from to the box around the Regional_Plot() inset; defaults to TRUE}

\item{Inset_Lines_Gap_Shading_Fill}{Colour of shaded space between lines from box around significant signals of interest from which inset expands from to the box around the Regional_Plot() inset; defaults to "grey70"}

\item{Inset_Lines_Gap_Shading_Alpha}{Opacity of shaded space between lines from box around significant signals of interest from which inset expands from to the box around the Regional_Plot() inset; defaults to 0.45}

\item{Peak_Highlight_To_Zero}{When Peak_Highlight_On is TRUE, ensure that the box drawn around significant signals of interest from which inset expands from reach to the baseline of background Single_Plot(); defaults to TRUE}

\item{Transparent_Regionals}{Ensure that Regional_Plot() insets are transparent, therefore allowing any overlapped elements on Single_Plot() background to be visible; defaults to FALSE}

\item{Preset_Regional_Width_Tweak}{Proportion of the full default Inset_Scale width of the Regional_Plot() panels to use when overlaying as insets allowing for further control in 2-Dimensions; defaults to 1}

\item{Peak_Highlight_Tweak_Y_Up}{Add/restrict spacing (matching to scale) along the maximum position on the Y-axis to the predefined Peak Highlight box around signal(s) of interest; defaults to 0.15}

\item{Peak_Highlight_Tweak_Y_Down}{Add/restrict spacing (matching to scale) along the minimum position on the Y-axis to the predefined Peak Highlight box around signal(s) of interest; defaults to 2}

\item{Peak_Highlight_Tweak_X}{Add/restrict spacing (matching to scale) along the midpoint on the X-axis to the predefined Peak Highlight box around signal(s) of interest; defaults to 0.0035}

\item{Inset_Lines_To_Zero}{When Inset_Lines_On is TRUE, ensure that lines from box around significant signals of interest from which inset expands from are also anchored to the baseline of the Single_Plot() background when Peak_Highlight_To_Zero is TRUE; defaults to TRUE}

\item{Transparent_Legends}{Ensure that Regional_Plot() legend insets are transparent, therefore allowing any overlapped elements on Single_Plot() background to be visible; defaults to FALSE}

\item{LD_Legend_On}{Show Auto_LD Legend derived from Regional_Plot object(s), when Auto_LD was TRUE; defaults to TRUE}

\item{Gene_Legend_On}{Show Gene_Legend, derived from Regional_Plot object(s); defaults to TRUE}

\item{Legend_Order}{When both LD_Legend_On & Gene_Legend_On are TRUE, decide the vertical order of the legends when using default spacing options; defaults to c("LD","Gene")}

\item{Legend_Gap_Frac}{Fixed vertical gap between the two legend boxes (as fraction of panel height); defaults to 0.04}

\item{LD_Legend_Horizontal}{Proportion of default sizing horizontal positioning to use when anchoring LD  key on Single_Plot() background; defaults to NULL, leading to max anchoring relative to default or specified LD_Legend_Preset}

\item{Gene_Legend_Horizontal}{Proportion of default sizing horizontal positioning to use when anchoring Gene Biotype key on Single_Plot() background; defaults to NULL, leading to max anchoring relative to default or specified Gene_Legend_Preset}

\item{Inset_Lines_Lower_Frac}{Proportion of the default Peak_Highlight_On box to draw the lower Inset_Line from; defaults to 0, allowing for full inset spacing}
}
\value{
Image of Manhattan Plot with inset Regional_Plot(s) is allocated to specified object and the resulting ggplot object can then be saved to an image
}
\description{
Customised Manhattan Plot with Inset Regional Plots
}
\examples{


Manhattan_Plot <- Single_Plot(Data = Household_Income_Sum_Stats, Label_Index = FALSE)

Regional_Plot <- Regional_Plot(Data = c(Household_Income_Sum_Stats, Household_Income_Sum_Stats),
                               Chromosomes = c(9,3),
                               Separation_Distance = 1e9,
                               Region_Window = c(3e5, 8e4),
                               Auto_LD = TRUE,
                               Genome_Build = "grch37",
                               Title_On = FALSE,
                               X_Axis_Title_Size = 100,
                               X_Axis_Title_Vjust = 0,
                               Y_Axis_Title_Vjust = 5.5,
                               Y_Axis_Title_Size = 100,
                               Y_Axis_Text_Size = 90,
                               Y_Axis_Title = "\n-log₁₀(P)",
                               Chromosome_Label_Size = 90,
                               Sig_Line_Width = 5,
                               Recombination_Line = TRUE,
                               Recombination_Line_Thickness = 7.5,
                               Recombination_Axis_Text_Size = 90,
                               Recombination_Axis_Title_Size= 100,
                               Recombination_Axis_Title_Vjust = 5.5,
                               Recombination_Axis_Title     = "\nRR (cM/Mb)",
                               Point_Size = 20,
                               Label_Size = 25,
                               Label_Height = 40,
                               Label_Angle  = 12.5,
                               Panel_Ratio  = "1:1",
                               Gene_Label_Vertical_Spacing = 0.25,
                               Gene_Track_Vertical_Spacing =  0.55,
                               LD_Legend_Title_Size = 20,
                               LD_Legend_Text_Size = 15,
                               Gene_Legend_Title_Size = 20,
                               Gene_Legend_Text_Size = 15,
                               LD_Legend_On = FALSE,
                               Gene_Legend_On = FALSE,
                               Manual_Decimal_Places  = c(3,2),
                               Target_Position_Breaks = 3,
                               Gene_Label_Buffer = 6.5,
                               Gene_Label_Size = 20,
                               Gene_Structure_Size_Exons = 24,
                               Gene_Structure_Size_Introns = 5,
                               Sense_Arrow_Body_Length = 0.02,
                               Sense_Arrow_Gene_Gap = 0.015,
                               Sense_Arrow_Head_Size = 7)


Annotated_Man <- Single_With_Regional_Plot(

  Single_Plot   = Manhattan_Plot,
  Regional_Plot = Regional_Plot,
  Inset_Chromosomes = c(3,9),
  Inset_Preset = c("left", "right"),
  Preset_Tweak_Y    = c(0.52, 0.4))




}
