% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_Tube_Alloys.R
\name{Tube_Alloys}
\alias{Tube_Alloys}
\title{Visualise GWAS Data}
\usage{
Tube_Alloys(
  Data = NULL,
  Chromosomes = 1,
  Phenos = NULL,
  Covars = NULL,
  Phenos_Covars = NULL,
  Pheno_Name = NULL
)
}
\arguments{
\item{Data}{These are the summary statistics (either file path(s), environmental object(s) or a variable(s) containing such contents) to be plotted; defaults to NULL}

\item{Chromosomes}{Manually specify which chromosomes from the summary statistics to run the chromosome sensitive functions on; defaults to NULL, leading to selective default filtering}

\item{Phenos}{Specify GWAS phenotype file, when Phenos_Covars is FALSE - used for GWAS model overview Forest_Plot(); defaults to NULL}

\item{Covars}{Specify GWAS covariate file, when Phenos_Covars is FALSE - used for GWAS model overview Forest_Plot(); defaults to NULL}

\item{Phenos_Covars}{Specify combined phenotype & covariate file, when Phenos & Covars are both FALSE - used for GWAS model overview Forest_Plot(); defaults to NULL}

\item{Pheno_Name}{When Phenos_Covars is TRUE, specify the name of the column containing the GWAS Phenotype - used for GWAS model overview Forest_Plot(); defaults to NULL}
}
\value{
Images of Manhattan Plot(s), Regional Plots(s) and Forest Plot(s) are allocated to specified object and the resulting ggplot and grob object(s) can then be saved to an image
}
\description{
Visualise GWAS Data
}
\examples{
Plots <- Tube_Alloys(Data = Intelligence_Sum_Stats, Phenos = Fake_PHENOS_Binary, Covars = Fake_COVARS, Chromosomes = 1 )

}
