% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate_Data.R
\name{Annotate_Data}
\alias{Annotate_Data}
\title{Title Use biomaRt to obtain RSID from COORDS if required before plotting}
\usage{
Annotate_Data(
  Data = Intelligence_Sum_Stats,
  Chromosome_Column = "CHROM",
  Position_Column = "GENPOS",
  SNP_ID_Column = "ID",
  PValue_Column = "P",
  Build = "HG38",
  Reference_Allele = "ALLELE0",
  Effect_Allele = "ALLELE1"
)
}
\arguments{
\item{Data}{This is the name of the data frame to annotate; defaults to Intelligence_Sum_Stats}

\item{Chromosome_Column}{Manually specify chromosome column; defaults to "CHROM"}

\item{Position_Column}{Manually specify chromosome position column; defaults to "GENPOS"}

\item{SNP_ID_Column}{Manually specify SNP ID column; defaults to "ID"}

\item{PValue_Column}{Manually specify P Value column; defaults to "P"}

\item{Build}{Genome reference to query biomaRt for ("HG38" or "HG19"); defaults to "HG38"}

\item{Reference_Allele}{The reference allele in the GWAS summary statistics; defaults to "ALLELE0"}

\item{Effect_Allele}{The alternate/test/effect allele in the GWAS summary statistics; defaults to "ALLELE1"}
}
\value{
Outputs the input dataframe with a column called Lab with
the corresponding RSIDs of the index SNP per chromosome
}
\description{
Title Use biomaRt to obtain RSID from COORDS if required before plotting
}
\examples{
 Labelled_Data <- Annotate_Data(Data = Intelligence_Sum_Stats,
           Chromosome_Column = "CHROM",
           Position_Column = "GENPOS", SNP_ID_Column = "ID",
           PValue_Column = "P", Build = "HG38")


}
