% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_Forest_Plot.R
\name{Forest_Plot}
\alias{Forest_Plot}
\title{Annotated, Customised Forest Plots}
\usage{
Forest_Plot(Data, ..., session = NULL)
}
\arguments{
\item{Data}{These are the summary statistics (either file path(s), environmental object(s) or a variable(s) containing such contents) to be plotted; defaults to NULL}

\item{Names}{These are a list of manual data set names to be applied to the legend text/key on the main plot - this also specifies the order down the page; defaults to NULL}

\item{Data_Set_Colours}{These are a list of manual data set colours allocated within the main plot - this also specifies the order down the page; defaults to NULL}

\item{Model_Reference}{Manually specify where the statistics to be used derive from (either from GWAS or Models); defaults to FALSE - not a model}

\item{Shapes}{These are a list of manual data set shapes allocated within the main plot - this also specifies the order down the page; defaults to NULL}

\item{Null_Line_Colour}{Colour of null line; defaults to "red"}

\item{Null_Line_Type}{Type of null line; defaults to "dashed"}

\item{X_Axis_Title}{X axis title; defaults to NULL}

\item{X_Axis_Title_Size}{X axis title size; defaults to 13}

\item{X_Axis_Label}{Display X axis title; defaults to TRUE}

\item{X_Axis_Separation}{Continuous value of the specific break separation to show on the X axis; defaults to NULL - leading to auto-calculation}

\item{Strip_Colour}{Colour of the strips highlighting the groups of covariates or SNPs under a certain category; defaults to NULL}

\item{Strips}{Display strips highlighting groups of interest; defaults to TRUE}

\item{Pre_Calculated_CIs}{Indicate that custom CIs are precalculated and available in datasets; defaults to FALSE}

\item{Legend_Title_Size}{Legend title size; defaults to 13}

\item{Legend_Text_Size}{Legend text size; defaults to 11}

\item{Legend_Title}{Title to display next to the legend; defaults to "Data"}

\item{Left_Title}{Title to display for the information in the grid text module on the left hand side of the plot; defaults to NULL}

\item{Match_Allele_Direction}{Match the allele directions to avoid effect direction flips; defaults to TRUE}

\item{Match_Allele_Study}{Data set used as reference point for effect allele direction; defaults to NULL - automatically assigned as first set}

\item{Selected_SNPs}{A list of SNPs to be shown in the forest plot; defaults to NULL - automatically selected for from the summary statistics data set}

\item{Selected_Covariates}{A list of covariates to be shown in the forest plot; defaults to NULL - automatically selected for from the munged model data set}

\item{X_Axis_Text_Resolution}{Number of decimal places to display on X axis text; defaults to NULL}

\item{Legend_On}{Display the main plot legend; defaults to TRUE}

\item{X_Axis_Text_Size}{Size of the X axis text labels; defaults to 11}

\item{Null_Buffer}{Units of space to avoid for X axis text labels around the NULL point; defaults to 0.00}

\item{Styles}{These are a list of styles to be applied to the grid text on the right hand side of the main plot - this also specifies the order down the page; defaults to NULL and accepts "underline" or "bold" as deviations}

\item{Missings}{Simulate shape positions for data points which may be missing from specific data sets plotted; defaults to TRUE}

\item{Null_Line_Thickness}{Manually specify the thickness of the corresponding Null Line; defaults to 0.5}

\item{Test_Statistic_Choice}{Manually specify which type of test statistic output is desired, OR or BETA; defaults to NULL - regardless of input type either output may be specified}

\item{Lead_Studies}{Manually specify which study's lead SNPs should be plotted when Selected_SNPs is NULL; defaults to NULL - leads to index SNPs from any set to be plotted}

\item{Separation_Distance}{The minimum genomic distance between significant regions of interest/lead SNPs to be defined as distinct peaks, defined by P_threshold value when Selected_SNPs is NULL; defaults to 1e6}

\item{P_threshold}{Maximum value of P (minimum LOG10P) allowed for a SNP to be classified as initiating detection as the lead SNP of a peak in the function, when Selected_SNPs is NULL; defaults to 5e-8}

\item{Chromosomes}{Manually specify which chromosomes from the summary statistics to run the function on, when Selected_SNPs is NULL; defaults to NULL - leads to all chromosomes being processed}

\item{Double_Label}{Display a second level of information under individual grid text layers on the left hand side of the main plot; defaults to FALSE}

\item{Shape_Size}{Manually specify the size of the Shapes added on main plot; defaults to 4}

\item{CI_Line_Height}{Manually specify the vertical length of the CI lines added on main plot; defaults to 0.35}

\item{CI_Line_Size}{Manually specify the thickness of the CI lines added on main plot; defaults to 0.4}

\item{Shape_Thickness}{Manually specify the thickness of the Shapes added on main plot; defaults to 1}

\item{Grid_Thickness}{Manually specify the thickness of the grid lines, when Grid_Lines_On is TRUE; defaults to 0.25}

\item{Grid_Lines_On}{Show grid lines with specified settings around left and right hand side annotations; defaults to TRUE}

\item{SE_Column}{Manually specify SE of Test Statistic column; defaults to NULL, leading to auto-detection}

\item{OR_Column}{Manually specify OR Test Statistic column; defaults to NULL, leading to auto-detection}

\item{Chromosome_Column}{Manually specify chromosome column; defaults to NULL, leading to auto-detection}

\item{BETA_Column}{Manually specify BETA Test Statistic column; defaults to NULL, leading to auto-detection}

\item{Position_Column}{Manually specify chromosome genomic position column; defaults to NULL, leading to auto-detection}

\item{SNP_ID_Column}{Manually specify SNP ID column; defaults to NULL, leading to auto-detection}

\item{PValue_Column}{Manually specify P Value column; defaults to NULL, leading to auto-detection}

\item{Reference_Allele_Column}{Manually specify reference allele column; defaults to NULL, leading to auto-detection}

\item{Effect_Allele_Column}{Manually specify effect allele column; defaults to NULL, leading to auto-detection}

\item{Upper_CI_Column}{Manually specify upper CI column, when Pre_Calculated_CIs is TRUE; defaults to NULL, leading to auto-detection}

\item{Lower_CI_Column}{Manually specify lower CI column, when Pre_Calculated_CIs is TRUE; defaults to NULL, leading to auto-detection}

\item{Left_Size}{Manually specify the size of the grid text on the left hand side of the main plot; defaults to 9}

\item{Right_Size}{Manually specify the size of the grid text on the right hand side of the main plot; defaults to 9}

\item{Right_Buffer}{Manually specify the size of space around the grid text on the right hand side of the main plot; defaults to 1}

\item{Left_Buffer}{Manually specify the size of space around the grid text on the left hand side of the main plot; defaults to 2}

\item{Axis_Buffer}{Manually specify the size of space around the axis end positions of the main plot; defaults to 0.02}

\item{Left_Fill}{Manually specify the colour of the grid text background on the left hand side of the main plot; defaults to "white"}

\item{Right_Fill}{Manually specify the colour of the grid text background on the right hand side of the main plot; defaults to "white"}

\item{Grid_Colour_Left}{Manually specify the colour of the grid lines on the left hand side of the main plot; defaults to "black"}

\item{Grid_Colour_Right}{Manually specify the colour of the grid lines on the right hand side of the main plot; defaults to "black"}

\item{CI_Label}{Allow the display of a column containing CI information in the grid text module on the right hand side of the plot; defaults to TRUE}

\item{BETA_Label}{Allow the display of a column containing BETA/OR information, depending on Test_Statistic_Choice in the grid text module on the right hand side of the plot; defaults to TRUE}

\item{SE_Label}{Allow the display of a column containing SE information in the grid text module on the right hand side of the plot; defaults to TRUE}

\item{BETA_SE_Label}{Allow the display of a column containing BETA & SE information in the grid text module on the right hand side of the plot; defaults to TRUE}

\item{BETA_CI_Label}{Allow the display of a column containing BETA & CI information in the grid text module on the right hand side of the plot; defaults to TRUE}

\item{P_Label}{Allow the display of a column containing P-Value information in the grid text module on the right hand side of the plot; defaults to TRUE}

\item{Special_P}{Format the column containing P-Value information in the grid text module on the right hand side of the plot when P_Label is TRUE and P_Label is in order() to display as * or NS; defaults to FALSE}

\item{CI_Label_Resolution}{Specify the number of decimal places of the column containing CI information in the grid text module on the right hand side of the plot; defaults to 3}

\item{BETA_Label_Resolution}{Specify the number of decimal places of the column containing BETA information in the grid text module on the right hand side of the plot; defaults to 3}

\item{SE_Label_Resolution}{Specify the number of decimal places of the column containing SE information in the grid text module on the right hand side of the plot; defaults to 3}

\item{P_Label_Resolution}{Specify the number of decimal places of the column containing P-Value information in the grid text module on the right hand side of the plot; defaults to 2}

\item{CI_Label_Title}{Specify the title of the column containing CI information in the grid text module on the right hand side of the plot; defaults to "CI"}

\item{BETA_Label_Title}{Specify the title of the column containing BETA information in the grid text module on the right hand side of the plot; defaults to NULL}

\item{SE_Label_Title}{Specify the title of the column containing SE information in the grid text module on the right hand side of the plot; defaults to "SE"}

\item{BETA_SE_Label_Title}{Specify the title of the column containing SE information in the grid text module on the right hand side of the plot; defaults to NULL}

\item{BETA_CI_Label_Title}{Specify the title of the column containing BETA & CI information in the grid text module on the right hand side of the plot; defaults to NULL}

\item{P_Label_Title}{Specify the title of the column containing P information in the grid text module on the right hand side of the plot; defaults to "P"}

\item{order}{Specify the order and selection of information containing columns (Labels) in the grid text module on the right hand side of the plot; defaults to c("BETA_SE_Label", "P_Label")}

\item{Verbose}{Prevent display of progress bar as function is running and instead show key milestone outputs/messages (mainly for debugging purposes); defaults to FALSE}

\item{Lab_Col}{Allocate the second level of information in the grid text module on the left hand side of the plot, when Double_Label is TRUE; defaults to "Real_ID"}
}
\value{
Image of Forest Plot is allocated to specified object and the resulting ggplot object can then be saved to an image
}
\description{
Annotated, Customised Forest Plots
}
\examples{
Forest_Plot <- Forest_Plot(Data = "Intelligence_Sum_Stats", Chromosomes = 1)

}
