% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_METASOFT_File_Gen.R
\name{METASOFT_File_Gen}
\alias{METASOFT_File_Gen}
\title{Munged summary statistics in preparation for METASOFT-like meta-analysis}
\usage{
METASOFT_File_Gen(..., session = NULL, Debug = FALSE)
}
\arguments{
\item{Match_Allele_Direction}{Match the allele directions to avoid effect direction flips; defaults to TRUE}

\item{Match_Allele_Study}{Data set used as reference point for effect allele direction; defaults to NULL - automatically assigned as first set}

\item{Output}{A selection of columns to isolate, in order, alternating by dataset; defaults to c("BETA", "SE") - standard METASOFT ordered format}

\item{Data}{These are the summary statistics (either file path(s), environmental object(s) or a variable(s) containing such contents) to be plotted; defaults to NULL}

\item{BETA_Column}{Manually specify BETA Test Statistic column; defaults to NULL, leading to auto-detection}

\item{Position_Column}{Manually specify chromosome genomic position column; defaults to NULL, leading to auto-detection}

\item{SNP_ID_Column}{Manually specify SNP ID column; defaults to NULL, leading to auto-detection}

\item{PValue_Column}{Manually specify P Value column; defaults to NULL, leading to auto-detection}

\item{Reference_Allele_Column}{Manually specify reference allele column; defaults to NULL, leading to auto-detection}

\item{Effect_Allele_Column}{Manually specify effect allele column; defaults to NULL, leading to auto-detection}

\item{SE_Column}{Manually specify SE of Test Statistic column; defaults to NULL, leading to auto-detection}
}
\value{
Munged dataframe amenable to METASOFT meta-analysis - only including SNPs present in backboned file
}
\description{
Munged summary statistics in preparation for METASOFT-like meta-analysis
}
\examples{
METASOFT <- METASOFT_File_Gen(Data = c("Intelligence_Sum_Stats", "Household_Income_Sum_Stats"))

}
