% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_Miami_Plot.R
\name{Miami_Plot}
\alias{Miami_Plot}
\title{Annotated, Customised Miami Plots}
\usage{
Miami_Plot(..., session = NULL)
}
\arguments{
\item{...}{Another shadow argument to detect inherited arguments passed and modified from the Single_Plot() function; defaults to NULL}

\item{session}{For safety, automatically blank session definitions; defaults to NULL}

\item{Top_Data}{These are the summary statistics (either file path(s),environmental object(s) or a variable(s) containing such contents) to be plotted on the top; defaults to NULL}

\item{Bottom_Data}{These are the summary statistics (either file path(s),environmental object(s) or a variable(s) containing such contents) to be plotted on the bottom; defaults to NULL}

\item{Chromosome_Column}{Manually specify chromosome column; defaults to NULL, leading to auto-detection}

\item{Position_Column}{Manually specify chromosome genomic position column; defaults to NULL, leading to auto-detection}

\item{SNP_ID_Column}{Manually specify SNP ID column; defaults to NULL, leading to auto-detection}

\item{PValue_Column}{Manually specify P Value column; defaults to NULL, leading to auto-detection}

\item{Reference_Allele_Column}{Manually specify reference allele column; defaults to NULL, leading to auto-detection}

\item{Effect_Allele_Column}{Manually specify effect allele column; defaults to NULL, leading to auto-detection}

\item{Verbose}{Prevent display of progress bar as function is running and instead show key milestone outputs/messages (mainly for debugging purposes); defaults to FALSE.}

\item{Adjust}{Shift shared X_Axis_Text down relative to the top plot, the larger the value the greater the shift; defaults to NULL and if not passed then calculated from expression mapped to Chromosome_Label_Size inherited from Single_Plot()}

\item{.dots}{Shadow argument to detect inherited arguments passed and modified from the Single_Plot() function; defaults to list()}
}
\value{
Image of Miami Plot(s) is allocated to specified object and the resulting combined grob object can then be saved to an image
}
\description{
Annotated, Customised Miami Plots
}
\examples{
Miami_Plot <- Miami_Plot(Top_Data = Household_Income_Sum_Stats,
                                   Bottom_Data = Intelligence_Sum_Stats)
}
