% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Munge.R
\name{Model_Munge}
\alias{Model_Munge}
\title{Title Process model output to data frame which can be used for plotting test statistics for different covariates}
\usage{
Model_Munge(Model_Object = Model, Model_Type = "glm")
}
\arguments{
\item{Model_Object}{This is the name of the base R model generated from running an lm() or glm(); defaults to Model}

\item{Model_Type}{This specifies the type of model run; defaults to "glm"}
}
\value{
Outputs the munged model data, ready for parsing to the Forest_Plot function
}
\description{
Title Process model output to data frame which can be used for plotting test statistics for different covariates
}
\examples{
ModelSumGLM <- Model_Munge(Model_Object = "Model", Model_Type = "glm")
          ModelSumLM <- Model_Munge(Model_Object = "Model2",  Model_Type = "lm")


}
