% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_Single_Plot.R
\name{Single_Plot}
\alias{Single_Plot}
\title{Annotated, Customised Manhattan Plots}
\usage{
Single_Plot(..., session = NULL)
}
\arguments{
\item{Data}{These are the summary statistics (either file path(s), environmental object(s) or a variable(s) containing such contents) to be plotted; defaults to NULL}

\item{Chromosome_Column}{Manually specify chromosome column; defaults to NULL, leading to auto-detection}

\item{Position_Column}{Manually specify chromosome genomic position column; defaults to NULL, leading to auto-detection}

\item{SNP_ID_Column}{Manually specify SNP ID column; defaults to NULL, leading to auto-detection}

\item{PValue_Column}{Manually specify P Value column; defaults to NULL, leading to auto-detection}

\item{Reference_Allele_Column}{Manually specify reference allele column; defaults to NULL, leading to auto-detection}

\item{Effect_Allele_Column}{Manually specify effect allele column; defaults to NULL, leading to auto-detection}

\item{Title}{Manually specify title to be displayed; defaults to NULL, leading to auto-detection of dataframe object/file name}

\item{Y_Axis_Title}{Manually specify Y axis title; defaults to "-log₁₀P"}

\item{X_Axis_Title}{Manually specify X axis title; defaults to NULL, leading to assignment as Chromosome when called solely in Single_Plot() but Genomic Position (Mb) within Regional_Plot() by default.}

\item{Title_Size}{Size of title; defaults to 35}

\item{Y_Axis_Text_Size}{Size of Y axis text; defaults to 30}

\item{Chromosome_Label_Size}{Size of chromosome numbers (X axis text); defaults to 30}

\item{Y_Axis_Title_Size}{Size of Y axis title; defaults to 35}

\item{X_Axis_Title_Size}{Size of X axis title; defaults to 35}

\item{Chromosome_Colours}{Alternating list of two colours allocated to data points (odd/even chromosome numbers); defaults to c("blue", "turquoise") - specifying the same colour twice will lead to uniform colouring.}

\item{Sig_Line_Colour}{Colour of significance threshold; defaults to "red"}

\item{Sig_Line_Type}{Type of significance threshold line; defaults to "dashed"}

\item{Sig_Threshold}{Value of significance threshold; defaults to standard Bonferroni correction of 5e-8}

\item{Sig_Line_Width}{Width of significance threshold line; defaults to 0.5}

\item{Point_Size}{Size of data points; defaults to 2.5 but defaults to 10 when called within Regional_Plot()}

\item{Label_Index}{Annotate the index SNPs with ID provided; defaults to TRUE}

\item{Label_Size}{Size of index labels if Label_Index is TRUE; defaults to 5 but defaults to 9 when called within Regional_Plot()}

\item{Label_Angle}{Angle of index labels if Label_Index is TRUE; defaults to 60 (degrees)}

\item{Label_Colour}{Colour of index labels if Label_Index is TRUE; defaults to "black"}

\item{Colour_Index}{Highlight the index SNP with a circular ring around the original point; defaults to TRUE}

\item{Colour_Of_Index}{Colour of index SNP circular highlight if Colour_Index is TRUE; defaults to "darkred"}

\item{Chromosome_Labels}{List of chromosomes to show index SNP labels on, given that Label_Index is TRUE; defaults to c(1:22, "X")}

\item{Chromosome_Index}{List of chromosomes to show index SNP circular highlight/colour on, given that Colour_Index is TRUE; defaults to c(1:22, "X")}

\item{Condense_Scale}{Log transform and condense scale above allotted break point to enhance visualisation of 'hits' with vastly different test statistics; defaults to TRUE}

\item{Break_Point}{Value of P where y axis begins to condense; defaults to 1e-10, guided by frequent max value observed in a small sample GWAS}

\item{Draft_Plot}{Specify whether a random subset of SNPs, the number of which is specified by Random_Selection, is plotted; defaults to FALSE}

\item{Random_Selection}{Manually select a random subset of SNPs to plot (more for internal debugging and testing), given that Draft_Plot is TRUE; defaults to 10,000}

\item{Chromosome_Label_Drops}{Manually specify which chromosome/X axis labels (as a list) should not be shown to prevent text crowding; defaults to NULL leading to c(21,22) being passed when X chromosome data is present and c(21) when only autosomes are present.}

\item{Interactive}{View the output as amenable to interactive plotly object, which can be used to inspect data, with key overlays (more for shiny App usage); defaults to FALSE}

\item{Index_Size}{Size of the outer ring diameter used to highlight the index SNPs, when Colour_Index is TRUE; defaults to NULL and if not manually passed is always scaled to 2X the passed or default Point_Size}

\item{Index_Thickness}{Thickness of the outer ring circumference used to highlight the index SNPs, when Colour_Index is TRUE; defaults to 1 but defaults to 4 when called within Regional_Plot()}

\item{Chromosome_Diamond}{List of chromosomes to show index SNP highlight/swapped diamond shape on, given that Diamond_Index is TRUE; defaults to c(1:22, "X")}

\item{Anchor_Label}{Which side of the point to anchor/position the index label to, when Label_Index is TRUE; defaults to "left"}

\item{Label_Height}{Spacing between the point and anchoring position of the index SNP label and the beginning of the label print area; defaults to 6 but defaults to 25 when called within Regional_Plot()}

\item{Diamond_Index}{Highlight the index SNP by swapping the standard point shape to a (different by default) coloured diamond; defaults to FALSE, unless called within Regional_Plot, then defaults to TRUE.}

\item{Colour_Of_Diamond}{Colour of diamond shape swapped for standard point shape ar index SNPs, when Diamond_Index is TRUE; defaults to "purple"}

\item{Diamond_Index_Size}{Size of the index SNP swapped diamond shape used to highlight the index SNPs, when Diamond_Index is TRUE; defaults to NULL and if not passed is scaled to 2X the according default or passed Point_Size}

\item{Lab}{Specify an additional info column with information to be labelled regarding Index SNPs, when Label_Index is TRUE; defaults to NULL leading to the values in the SNP ID column being used (Coords/RS codes) by default}

\item{Verbose}{Prevent display of progress bar as function is running and instead show key milestone outputs/messages (mainly for debugging purposes); defaults to FALSE.}

\item{Auto_Lab}{Allow Annotate_Data() to run on the data frame(s) provided so Lab can have RSIDs assigned, if only coordinates exist; defaults to FALSE}

\item{Genome_Build}{Reference genome to base required RSID annotations around; defaults to grch38}
}
\value{
Image of Manhattan Plot(s) is allocated to specified object and the resulting ggplot object can then be saved to an image
}
\description{
Annotated, Customised Manhattan Plots
}
\examples{
 Manhattan_Plot <- Single_Plot(Data = Intelligence_Sum_Stats)

 Manhattan_Plot <- Single_Plot(Data = Intelligence_Sum_Stats)
}
